FI-1161 refactor: move all pack-specific types to `autotests/configurator`
