+++ README.md
@@ -173,7 +173,7 @@ This pack might look like this:
 ```ts
 import {pack as allTestsPack} from './allTests';
 
-import type {Pack} from 'autotests/types/pack';
+import type {Pack} from 'autotests/configurator';
 
 /**
  * Pack from .gitignore for local development.
