+++ autotests/configurator/index.ts
@@ -6,3 +6,21 @@ export {mapLogPayloadInConsole} from './mapLogPayloadInConsole';
 export {mapLogPayloadInLogFile} from './mapLogPayloadInLogFile';
 export {mapLogPayloadInReport} from './mapLogPayloadInReport';
 export {skipTests} from './skipTests';
+export type {
+  DoAfterPack,
+  DoBeforePack,
+  FilterTestsIntoPack,
+  GetFullPackConfig,
+  GetLogContext,
+  GetMainTestRunParams,
+  GetTestRunHash,
+  IsTestSkipped,
+  LiteReport,
+  MapBackendResponseErrorToLog,
+  MapBackendResponseToLog,
+  MapLogPayloadInConsole,
+  MapLogPayloadInLogFile,
+  MapLogPayloadInReport,
+  Pack,
+  TestMeta,
+} from './types';
