+++ autotests/configurator/types/customPackProperties.ts
