+++ autotests/configurator/types/customReportProperties.ts
