+++ autotests/configurator/types/index.ts
@@ -1,0 +1,18 @@
+export type {
+  DoAfterPack,
+  DoBeforePack,
+  FilterTestsIntoPack,
+  GetFullPackConfig,
+  GetLogContext,
+  GetMainTestRunParams,
+  GetTestRunHash,
+  IsTestSkipped,
+  LiteReport,
+  MapBackendResponseErrorToLog,
+  MapBackendResponseToLog,
+  MapLogPayloadInConsole,
+  MapLogPayloadInLogFile,
+  MapLogPayloadInReport,
+  Pack,
+} from './packSpecific';
+export type {TestMeta} from './testMeta';
