+++ autotests/configurator/types/pack.ts
