+++ autotests/configurator/types/testMeta.ts
