+++ autotests/types/index.ts
@@ -30,10 +30,3 @@ export type {
   UserPhone,
   UserWorker,
 } from './entities';
-export type {
-  GetLogContext,
-  GetMainTestRunParams,
-  GetTestRunHash,
-  IsTestSkipped,
-} from './packSpecific';
-export type {TestMeta} from './testMeta';
