+++ autotests/utils/getFullPackConfig.ts
@@ -1,6 +1,6 @@
 import {untypedGetFullPackConfig} from 'e2ed/utils';
 
-import type {GetFullPackConfig} from 'autotests/types/packSpecific';
+import type {GetFullPackConfig} from 'autotests/configurator';
 
 /**
  * Get full pack configuration object.
